Use size_t / ssize_t
The library uses the sendmsg / recv / splice / vmsplice system calls for
the communication with the kernel. Those are defined to use
size_t/ssize_t variable types. The library on the other hand used
uint32_t / int32_t which does not comply and caused a lot of warnings
with -Wconversion.

This patch started by changing the variable types used to invoke the
system call to ssize_t / size_t. This change metastized into all code
parts of the library to use the updated data types.

The change now allows the enabling of -Wconversion and limit the number
of type casts.

+++ apps/app-internal.c
@@ -225,10 +225,10 @@ int hex2bin_alloc(const char *hex, uint32_t hexlen,
 	return 0;
 }
 
-int read_complete(int fd, uint8_t *buf, uint32_t buflen)
+ssize_t read_complete(int fd, uint8_t *buf, size_t buflen)
 {
 	ssize_t ret;
-	int rc = 0;
+	ssize_t rc = 0;
 
 	if (buflen > INT_MAX)
 		return -EINVAL;
@@ -236,7 +236,7 @@ int read_complete(int fd, uint8_t *buf, uint32_t buflen)
 	do {
 		ret = read(fd, buf, buflen);
 		if (0 < ret) {
-			buflen -= ret;
+			buflen -= (size_t)ret;
 			buf += ret;
 		}
 		rc += ret;
